Remove Tailwind
Tailwind is nice for prototyping. However, we'll now start to write a
lot more custom CSS, and as we do that, our custom CSS will interact
with Tailwind's CSS, and it will be difficult to tell which styles are
coming from where, which will add work. This is a good time to remove
Tailwind, since we're redoing most of the CSS anyway.