Update sorting strategy
When Javascript is off, companies are sorted by descending number of
devs. When Javascript is on, the first company is always the biggest
one, then each subsequent company in the sorted list has a 50% chance of
getting skipped, and this is carried out until 5 companies are chosen.
This means that predominantly large companies will be shown, but
everyone has a chance to get shown.