Clarify annual report dates
We've previously used confusing wording for this, eg by saying that
members should publish reports “at least” yearly. This gets confusing
because, if members published more frequently than yearly, the reports
themselves would not be yearly reports, so we would either be unable to
compare them, or would have to somehow pro-rate or merge them.

This new description clarifies that reports are yearly, and must all end
on the same day of the year, but can be published at any time, eg
earlier in the year.