chore: update members/bolt.json
Hey! 

As we don't know yet if the distribution of the 100K fund will drip into the next year and who are the recipients, I think what we could do now is add the 2025 entry with the projected 75K for the currently announced monthly sponsorings in the [blog post](https://blog.stackblitz.com/posts/bolt-100k-oss-fund/). The entry can then be updated with a report later in the year or at the beginning of 26.